Run database tests against MySQL, not SQLite
